backend/requirements.txt
flask==1.1.1
twilio==6.35.5
firebase-admin==4.0.0
spotipy==2.9.0
sanic==19.12.0
google-cloud-tasks==1.5.0
googleapis-common-protos==1.6.0
mutagen==1.45.1
phonenumbers==8.12.1
aiohttp[speedups]
